+++ integration_tests/suite/test_applications.py
@@ -5,7 +5,9 @@
 from hamcrest import (
     assert_that,
     contains,
+    contains_inanyorder,
     empty,
+    equal_to,
     has_entries,
     has_items,
     has_properties,
@@ -616,6 +618,369 @@ class TestApplicationHold(BaseApplicationTestCase):
                 )
             )
         )
+
+
+class TestApplicationSnoop(BaseApplicationTestCase):
+
+    def setUp(self):
+        super(TestApplicationSnoop, self).setUp()
+        self.app_uuid = self.no_node_app_uuid
+        self.caller_channel = self.call_app(self.no_node_app_uuid)
+        node = self.ctid_ng.application_new_node(
+            self.app_uuid,
+            calls=[self.caller_channel.id],
+        ).json()
+        self.ctid_ng.application_new_node_call(
+            self.app_uuid,
+            node['uuid'],
+            'local',
+            'recipient_autoanswer',
+        )
+
+    def test_snoop_created_event(self):
+        supervisor_channel = self.ctid_ng.application_new_call(
+            self.app_uuid,
+            'local',
+            'recipient_autoanswer',
+        ).json()
+
+        routing_key = 'applications.{uuid}.snoops.#'.format(uuid=self.app_uuid)
+        event_accumulator = self.bus.accumulator(routing_key)
+
+        whisper_mode = 'both'
+        snoop = self.ctid_ng.application_call_snoop(
+            self.app_uuid,
+            self.caller_channel.id,
+            supervisor_channel['id'],
+            whisper_mode,
+        ).json()
+
+        def event_received():
+            events = event_accumulator.accumulate()
+            assert_that(
+                events,
+                contains(
+                    has_entries(
+                        name='application_snoop_created',
+                        data=has_entries(
+                            application_uuid=self.app_uuid,
+                            snoop=has_entries(
+                                uuid=snoop['uuid'],
+                                snooped_call_id=self.caller_channel.id,
+                                snooping_call_id=supervisor_channel['id'],
+                                whisper_mode=whisper_mode,
+                            )
+                        )
+                    )
+                )
+            )
+
+        until.assert_(event_received, tries=3)
+
+    def test_snoop_deleted_event(self):
+        supervisor_channel = self.ctid_ng.application_new_call(
+            self.app_uuid,
+            'local',
+            'recipient_autoanswer',
+        ).json()
+
+        snoop = self.ctid_ng.application_call_snoop(
+            self.app_uuid,
+            self.caller_channel.id,
+            supervisor_channel['id'],
+            'both',
+        ).json()
+
+        routing_key = 'applications.{uuid}.snoops.#'.format(uuid=self.app_uuid)
+        event_accumulator = self.bus.accumulator(routing_key)
+
+        self.ctid_ng.application_delete_snoop(self.app_uuid, snoop['uuid'])
+
+        def event_received():
+            events = event_accumulator.accumulate()
+            assert_that(
+                events,
+                contains(
+                    has_entries(
+                        name='application_snoop_deleted',
+                        data=has_entries(
+                            application_uuid=self.app_uuid,
+                            snoop=has_entries(uuid=snoop['uuid']),
+                        )
+                    )
+                )
+            )
+
+        until.assert_(event_received, tries=3)
+
+    def test_snoop_updated_event(self):
+        supervisor_channel = self.ctid_ng.application_new_call(
+            self.app_uuid,
+            'local',
+            'recipient_autoanswer',
+        ).json()
+
+        snoop = self.ctid_ng.application_call_snoop(
+            self.app_uuid,
+            self.caller_channel.id,
+            supervisor_channel['id'],
+            'both',
+        ).json()
+
+        routing_key = 'applications.{uuid}.snoops.#'.format(uuid=self.app_uuid)
+        event_accumulator = self.bus.accumulator(routing_key)
+
+        whisper_mode = 'in'
+        self.ctid_ng.application_edit_snoop(
+            self.app_uuid,
+            snoop['uuid'],
+            whisper_mode,
+        )
+
+        def event_received():
+            events = event_accumulator.accumulate()
+            assert_that(
+                events,
+                contains(
+                    has_entries(
+                        name='application_snoop_updated',
+                        data=has_entries(
+                            application_uuid=self.app_uuid,
+                            snoop=has_entries(
+                                uuid=snoop['uuid'],
+                                snooped_call_id=self.caller_channel.id,
+                                snooping_call_id=supervisor_channel['id'],
+                                whisper_mode=whisper_mode,
+                            )
+                        )
+                    )
+                )
+            )
+
+        until.assert_(event_received, tries=3)
+
+    def test_delete(self):
+        supervisor_1_channel = self.ctid_ng.application_new_call(
+            self.app_uuid,
+            'local',
+            'recipient_autoanswer',
+        ).json()
+        supervisor_2_channel = self.ctid_ng.application_new_call(
+            self.app_uuid,
+            'local',
+            'recipient_autoanswer',
+        ).json()
+
+        snoop_1 = self.ctid_ng.application_call_snoop(
+            self.app_uuid,
+            self.caller_channel.id,
+            supervisor_1_channel['id'],
+            'both',
+        ).json()
+
+        snoop_2 = self.ctid_ng.application_call_snoop(
+            self.app_uuid,
+            self.caller_channel.id,
+            supervisor_2_channel['id'],
+            'both',
+        ).json()
+
+        result = self.ctid_ng.application_delete_snoop(self.app_uuid, snoop_2['uuid'])
+        assert_that(result, has_properties(status_code=204))
+        assert_that(
+            self.ctid_ng.application_list_snoops(self.app_uuid).json(),
+            has_entries(
+                items=contains_inanyorder(
+                    snoop_1,
+                )
+            )
+        )
+
+        result = self.ctid_ng.application_delete_snoop(self.app_uuid, snoop_2['uuid'])
+        assert_that(result, has_properties(status_code=404))
+
+        result = self.ctid_ng.application_delete_snoop(self.unknown_uuid, snoop_2['uuid'])
+        assert_that(result, has_properties(status_code=404))
+
+    def test_list(self):
+        result = self.ctid_ng.application_list_snoops(self.app_uuid)
+        assert_that(
+            result.json(),
+            has_entries(items=empty())
+        )
+
+        result = self.ctid_ng.application_list_snoops(self.unknown_uuid)
+        assert_that(result, has_properties(status_code=404))
+
+        supervisor_1_channel = self.ctid_ng.application_new_call(
+            self.app_uuid,
+            'local',
+            'recipient_autoanswer',
+        ).json()
+        supervisor_2_channel = self.ctid_ng.application_new_call(
+            self.app_uuid,
+            'local',
+            'recipient_autoanswer',
+        ).json()
+
+        snoop_1 = self.ctid_ng.application_call_snoop(
+            self.app_uuid,
+            self.caller_channel.id,
+            supervisor_1_channel['id'],
+            'both',
+        ).json()
+
+        snoop_2 = self.ctid_ng.application_call_snoop(
+            self.app_uuid,
+            self.caller_channel.id,
+            supervisor_2_channel['id'],
+            'both',
+        ).json()
+
+        result = self.ctid_ng.application_list_snoops(self.app_uuid)
+        assert_that(
+            result.json(),
+            has_entries(
+                items=contains_inanyorder(
+                    snoop_1,
+                    snoop_2,
+                )
+            )
+        )
+
+    def test_get(self):
+        supervisor_channel = self.ctid_ng.application_new_call(
+            self.app_uuid,
+            'local',
+            'recipient_autoanswer',
+        ).json()
+        snoop = self.ctid_ng.application_call_snoop(
+            self.app_uuid,
+            self.caller_channel.id,
+            supervisor_channel['id'],
+            'both',
+        ).json()
+
+        result = self.ctid_ng.application_get_snoop(self.app_uuid, snoop['uuid'])
+        assert_that(result.json(), equal_to(snoop))
+
+        result = self.ctid_ng.application_get_snoop(self.unknown_uuid, snoop['uuid'])
+        assert_that(result, has_properties(status_code=404))
+
+        result = self.ctid_ng.application_get_snoop(self.app_uuid, self.unknown_uuid)
+        assert_that(result, has_properties(status_code=404))
+
+    def test_post_snoop(self):
+        unrelated_channel = self.call_app(self.node_app_uuid)
+        supervisor_channel = self.ctid_ng.application_new_call(
+            self.app_uuid,
+            'local',
+            'recipient_autoanswer',
+        ).json()
+
+        result = self.ctid_ng.application_call_snoop(
+            self.unknown_uuid,
+            self.caller_channel.id,
+            supervisor_channel['id'],
+            'both',
+        )
+        assert_that(result, has_properties(status_code=404))
+
+        result = self.ctid_ng.application_call_snoop(
+            self.app_uuid,
+            unrelated_channel.id,
+            supervisor_channel['id'],
+            'both',
+        )
+        assert_that(result, has_properties(status_code=404))
+
+        result = self.ctid_ng.application_call_snoop(
+            self.app_uuid,
+            self.caller_channel.id,
+            unrelated_channel.id,
+            'both',
+        )
+        assert_that(result, has_properties(status_code=400))
+
+        invalid_whisper_mode = [
+            'foobar',
+            'In',
+            False,
+            True,
+            42,
+            [],
+            {},
+        ]
+        for whisper_mode in invalid_whisper_mode:
+            result = self.ctid_ng.application_call_snoop(
+                self.app_uuid,
+                self.caller_channel.id,
+                supervisor_channel['id'],
+                whisper_mode,
+            )
+            assert_that(result, has_properties(status_code=400), whisper_mode)
+
+        result = self.ctid_ng.application_call_snoop(
+            self.app_uuid,
+            self.caller_channel.id,
+            supervisor_channel['id'],
+            'both',
+        )
+
+        assert_that(
+            result.json(),
+            has_entries(
+                uuid=uuid_(),
+                whisper_mode='both',
+                snooped_call_id=self.caller_channel.id,
+                snooping_call_id=supervisor_channel['id'],
+            )
+        )
+
+    def test_put(self):
+        supervisor_channel = self.ctid_ng.application_new_call(
+            self.app_uuid,
+            'local',
+            'recipient_autoanswer',
+        ).json()
+
+        snoop = self.ctid_ng.application_call_snoop(
+            self.app_uuid,
+            self.caller_channel.id,
+            supervisor_channel['id'],
+            'both',
+        ).json()
+
+        result = self.ctid_ng.application_edit_snoop(
+            self.unknown_uuid,
+            snoop['uuid'],
+            'in',
+        )
+        assert_that(result, has_properties(status_code=404))
+
+        result = self.ctid_ng.application_edit_snoop(
+            self.app_uuid,
+            self.unknown_uuid,
+            'in',
+        )
+        assert_that(result, has_properties(status_code=404))
+
+        invalid_whisper_mode = [
+            'foobar',
+            'In',
+            False,
+            True,
+            42,
+            [],
+            {},
+        ]
+        for whisper_mode in invalid_whisper_mode:
+            result = self.ctid_ng.application_edit_snoop(
+                self.app_uuid,
+                snoop['uuid'],
+                whisper_mode,
+            )
+            assert_that(result, has_properties(status_code=400), whisper_mode)
 
 
 class TestApplicationMoh(BaseApplicationTestCase):